Arreglo de bugs e implementacion de rondas
Arregle algunos errores que salian en la consola e implemente el sistema de rondas, con scripts colocados en la carpeta Rounds, tambien sincronizados con HUDController.

+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -62,7 +62,7 @@ public class PlayerMovement : MonoBehaviour
     {
         if (thirdPersonAnimator != null)
         {
-            thirdPersonAnimator.SetFloat("old_pistol", 0f);
+            thirdPersonAnimator.SetInteger("old_pistol", 0);
         }
     }
